pyproject.toml
[tool.poetry]
name = "kitchen"
version = "0.1.0"
description = ""
authors = ["Fernando Crozetta <fernando@czetta.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
psycopg2-binary = "^2.9.5"
mysql-connector-python = "^8.0.32"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pylint = "^2.16.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
